optimiziing code

+++ service/service_order/src/main/java/com/btwl/eduorder/controller/OrderController.java
@@ -56,7 +56,7 @@ public class OrderController {
     wrapper.eq("course_id", courseId);
     wrapper.eq("member_id", memberId);
     wrapper.eq("status", 1);//支付状态 1代表已经支付
-    int count = orderService.count(wrapper);
+    int count = (int) orderService.count(wrapper);
     if (count > 0) { //已经支付
       return true;
     } else {
